AG-14332 Move duplicate code into shared util-file
